Extract React tests from pretty-format-test.js (#3180)
* Extract React tests from pretty-format-test.js

* Mock console in one place only
